requirements.txt
# AureonCore - Backend híbrido (Aureon + Runa)
# Python 3.11+

# Core Framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0
pydantic-settings>=2.7.0

# Async HTTP
httpx>=0.28.0
aiohttp>=3.11.0

# Database
supabase>=2.11.0

# AI Providers
google-generativeai>=0.8.0
groq>=0.12.0
mistralai>=1.2.0

# Security
cryptography>=44.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Utilities
python-dotenv>=1.0.0
python-multipart>=0.0.15
pypdf>=4.0.0
python-docx>=1.1.0
